feat(NonSlidingPiece): NonSlidingPiece 추상 클래스 생성
