+++ src/main/java/chess/domain/piece/Knight.java
@@ -3,21 +3,11 @@ package chess.domain.piece;
 import chess.domain.color.Color;
 import java.util.Set;
 
-public class Knight extends Piece {
-    private static Set<Direction> directions = Direction.getKnightDirection();
+public class Knight extends NonSlidingPiece {
+    private static Set<Direction> DIRECTIONS = Direction.getKnightDirection();
 
     public Knight(Position position, Color color) {
-        super(position, color);
-    }
-
-    @Override
-    public Set<Position> findMovablePositions(Position destination) {
-        Set<Position> movable = position.findMovablePositions(directions);
-
-        if (movable.contains(destination)) {
-            return Set.of(destination);
-        }
-        throw new IllegalArgumentException("이동할 수 없습니다.");
+        super(position, color, DIRECTIONS);
     }
 
     @Override
