+++ src/main/java/chess/domain/piece/NonSlidingPiece.java
@@ -1,0 +1,24 @@
+package chess.domain.piece;
+
+import chess.domain.color.Color;
+import java.util.Set;
+
+public abstract class NonSlidingPiece extends Piece {
+    private final Set<Direction> directions;
+
+    public NonSlidingPiece(Position position, Color color, Set<Direction> directions) {
+        super(position, color);
+        this.directions = directions;
+    }
+
+    public Set<Position> findMovablePositions(Position destination) {
+        Set<Position> movable = position.findMovablePositions(directions);
+
+        if (movable.contains(destination)) {
+            return Set.of(destination);
+        }
+        throw new IllegalArgumentException("이동할 수 없습니다.");
+    }
+
+    public abstract PieceType pieceType();
+}
